pyproject.toml
[tool.poetry]
name = "abandonauth"
version = "0.1.0"
description = ""
authors = ["GDWR <gregory.dwr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
toml = "^0.10.2"
redis = {version = ">=4.2.0rc1", extras = ["hiredis"]}
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^0.20.0"
classy-config = "^0.3.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-annotations = "^2.9.0"
flake8-bandit = "^3.0.0"
flake8-bugbear = "^22.7.1"
flake8-docstrings = "^1.6.0"
flake8-import-order = "^0.18.1"
flake8-string-format = "^0.3.0"
pep8-naming = "^0.13.1"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "abandonauth._run:main"
